refactor(desktop): mac-path 并入 os-startup-tweaks，按平台拆分初始化
启动微调收口为单一 OS/平台域并更名（process-tweaks → os-startup-tweaks，applyProcessStartupTweaks →
applyOsStartupTweaks）：
- mac-path.ts 并入：fixMacPath 改为内部 augmentMacPath，不再单独导出 / 不再返回值（PATH 补全详情日志
  去除——与其余静默微调一致）。
- 按平台拆出 applyWindowsStartupTweaks（chcp）/ applyMacStartupTweaks（mock-keychain + PATH），平台
  判断只在 applyOsStartupTweaks 的分支里做一次（去掉各函数内的冗余平台检查）；PYTHONDONTWRITEBYTECODE
  为跨平台项留在外层。
- index 同步：移除 fixMacPath 调用与 macOS PATH 日志，main 仅 applyOsStartupTweaks()。

纯结构 / 封装、行为不变（仅去一条 PATH 补全的诊断日志）；desktop typecheck/lint/build 全绿。
